## Overview
**Issue Type**

- [ ] Bug
- [x] Feature
- [ ] Tech Debt

**Description**
Completed and styled the homepage which allows the user to click and navigate to the login page. Also added a favicon for the browser tab. Added new font mynerve to be able to use in the project. The logo in the header will try to navigate back to the homepage if you click it. If you are logged in it will stay on the current page.